node test: Extract compose_video tests.
