+++ frontend_tests/node_tests/compose.js
@@ -11,8 +11,6 @@ const {set_global, zrequire} = require("../zjsunit/namespace");
 const {run_test} = require("../zjsunit/test");
 const $ = require("../zjsunit/zjquery");
 
-const events = require("./lib/events");
-
 const noop = () => {};
 
 set_global("DOMParser", new JSDOM().window.DOMParser);
@@ -101,7 +99,6 @@ MockDate.set(new Date(fake_now * 1000));
 
 rewiremock.enable();
 
-const compose_ui = zrequire("compose_ui");
 const peer_data = zrequire("peer_data");
 const util = zrequire("util");
 const rtl = zrequire("rtl");
@@ -112,21 +109,9 @@ const compose_pm_pill = zrequire("compose_pm_pill");
 const echo = zrequire("echo");
 const compose = zrequire("compose");
 const upload = zrequire("upload");
-const server_events_dispatch = zrequire("server_events_dispatch");
 const settings_config = zrequire("settings_config");
 
 people.small_avatar_url_for_person = () => "http://example.com/example.png";
-
-function stub_out_video_calls() {
-    const elem = $("#below-compose-content .video_link");
-    elem.toggle = (show) => {
-        if (show) {
-            elem.show();
-        } else {
-            elem.hide();
-        }
-    };
-}
 
 function reset_jquery() {
     // Avoid leaks.
@@ -1037,10 +1022,12 @@ test_ui("warn_if_private_stream_is_linked", () => {
     }
 });
 
-test_ui("initialize", () => {
+test_ui("initialize", (override) => {
     // In this test we mostly do the setup stuff in addition to testing the
     // normal workflow of the function. All the tests for the on functions are
     // done in subsequent tests directly below this test.
+
+    override(compose, "compute_show_video_chat_button", () => false);
 
     let resize_watch_manual_resize_checked = false;
     resize.watch_manual_resize = (elem) => {
@@ -1071,29 +1058,6 @@ test_ui("initialize", () => {
         };
     });
 
-    page_params.realm_available_video_chat_providers = {
-        disabled: {
-            id: 0,
-            name: "disabled",
-        },
-        jitsi_meet: {
-            id: 1,
-            name: "Jitsi Meet",
-        },
-        zoom: {
-            id: 3,
-            name: "Zoom",
-        },
-        big_blue_button: {
-            id: 4,
-            name: "Big Blue Button",
-        },
-    };
-
-    page_params.realm_video_chat_provider =
-        page_params.realm_available_video_chat_providers.disabled.id;
-
-    stub_out_video_calls();
     compose.initialize();
 
     assert(resize_watch_manual_resize_checked);
@@ -1110,7 +1074,6 @@ test_ui("initialize", () => {
         page_params.narrow = true;
 
         reset_jquery();
-        stub_out_video_calls();
         set_up_compose_start_mock({});
 
         compose.initialize();
@@ -1122,7 +1085,6 @@ test_ui("initialize", () => {
         page_params.narrow_topic = "testing";
 
         reset_jquery();
-        stub_out_video_calls();
         set_up_compose_start_mock({topic: "testing"});
 
         compose.initialize();
@@ -1134,7 +1096,6 @@ test_ui("initialize", () => {
         $("#compose-send-button").prop("disabled", true);
 
         reset_jquery();
-        stub_out_video_calls();
         compose.initialize();
 
         compose.abort_xhr();
@@ -1555,148 +1516,6 @@ test_ui("on_events", () => {
         assert(compose_file_input_clicked);
     })();
 
-    (function test_no_provider_video_link_compose_clicked() {
-        let called = false;
-
-        const textarea = $.create("target-stub");
-
-        const ev = {
-            preventDefault: noop,
-            target: {
-                to_$: () => textarea,
-            },
-        };
-
-        compose_ui.__Rewire__("insert_syntax_and_focus", () => {
-            called = true;
-        });
-
-        const handler = $("body").get_on_handler("click", ".video_link");
-        $("#compose-textarea").val("");
-
-        handler(ev);
-        assert(!called);
-    })();
-
-    (function test_jitsi_video_link_compose_clicked() {
-        let syntax_to_insert;
-        let called = false;
-
-        const textarea = $.create("jitsi-target-stub");
-
-        const ev = {
-            preventDefault: noop,
-            target: {
-                to_$: () => textarea,
-            },
-        };
-
-        compose_ui.__Rewire__("insert_syntax_and_focus", (syntax) => {
-            syntax_to_insert = syntax;
-            called = true;
-        });
-
-        const handler = $("body").get_on_handler("click", ".video_link");
-        $("#compose-textarea").val("");
-
-        page_params.realm_video_chat_provider =
-            page_params.realm_available_video_chat_providers.jitsi_meet.id;
-
-        page_params.jitsi_server_url = null;
-        handler(ev);
-        assert(!called);
-
-        page_params.jitsi_server_url = "https://meet.jit.si";
-        handler(ev);
-        // video link ids consist of 15 random digits
-        const video_link_regex = /\[translated: Click to join video call]\(https:\/\/meet.jit.si\/\d{15}\)/;
-        assert(called);
-        assert.match(syntax_to_insert, video_link_regex);
-    })();
-
-    (function test_zoom_video_link_compose_clicked() {
-        let syntax_to_insert;
-        let called = false;
-
-        const textarea = $.create("zoom-target-stub");
-
-        const ev = {
-            preventDefault: noop,
-            target: {
-                to_$: () => textarea,
-            },
-        };
-
-        compose_ui.__Rewire__("insert_syntax_and_focus", (syntax) => {
-            syntax_to_insert = syntax;
-            called = true;
-        });
-
-        const handler = $("body").get_on_handler("click", ".video_link");
-        $("#compose-textarea").val("");
-
-        page_params.realm_video_chat_provider =
-            page_params.realm_available_video_chat_providers.zoom.id;
-        page_params.has_zoom_token = false;
-
-        window.open = (url) => {
-            assert(url.endsWith("/calls/zoom/register"));
-
-            // The event here has value=true.  We keep it in events.js to
-            // allow our tooling to verify its schema.
-            server_events_dispatch.dispatch_normal_event(events.fixtures.has_zoom_token);
-        };
-
-        channel.post = (payload) => {
-            assert.equal(payload.url, "/json/calls/zoom/create");
-            payload.success({url: "example.zoom.com"});
-            return {abort: () => {}};
-        };
-
-        handler(ev);
-        const video_link_regex = /\[translated: Click to join video call]\(example\.zoom\.com\)/;
-        assert(called);
-        assert.match(syntax_to_insert, video_link_regex);
-    })();
-
-    (function test_bbb_video_link_compose_clicked() {
-        let syntax_to_insert;
-        let called = false;
-
-        const textarea = $.create("bbb-target-stub");
-
-        const ev = {
-            preventDefault: noop,
-            target: {
-                to_$: () => textarea,
-            },
-        };
-
-        compose_ui.__Rewire__("insert_syntax_and_focus", (syntax) => {
-            syntax_to_insert = syntax;
-            called = true;
-        });
-
-        const handler = $("body").get_on_handler("click", ".video_link");
-        $("#compose-textarea").val("");
-
-        page_params.realm_video_chat_provider =
-            page_params.realm_available_video_chat_providers.big_blue_button.id;
-
-        channel.get = (options) => {
-            assert(options.url === "/json/calls/bigbluebutton/create");
-            options.success({
-                url:
-                    "/calls/bigbluebutton/join?meeting_id=%22zulip-1%22&password=%22AAAAAAAAAA%22&checksum=%2232702220bff2a22a44aee72e96cfdb4c4091752e%22",
-            });
-        };
-
-        handler(ev);
-        const video_link_regex = /\[translated: Click to join video call]\(\/calls\/bigbluebutton\/join\?meeting_id=%22zulip-1%22&password=%22AAAAAAAAAA%22&checksum=%2232702220bff2a22a44aee72e96cfdb4c4091752e%22\)/;
-        assert(called);
-        assert.match(syntax_to_insert, video_link_regex);
-    })();
-
     (function test_markdown_preview_compose_clicked() {
         // Tests setup
         function setup_visibilities() {
@@ -1907,32 +1726,6 @@ test_ui("narrow_button_titles", () => {
     compose.update_closed_compose_buttons_for_stream();
     assert.equal($("#left_bar_compose_stream_button_big").text(), i18n.t("New topic"));
     assert.equal($("#left_bar_compose_private_button_big").text(), i18n.t("New private message"));
-});
-
-test_ui("test_video_chat_button_toggle", () => {
-    reset_jquery();
-    stub_out_video_calls();
-
-    page_params.realm_video_chat_provider =
-        page_params.realm_available_video_chat_providers.disabled.id;
-    compose.initialize();
-    assert.equal($("#below-compose-content .video_link").visible(), false);
-
-    reset_jquery();
-    stub_out_video_calls();
-    page_params.realm_video_chat_provider =
-        page_params.realm_available_video_chat_providers.jitsi_meet.id;
-    page_params.jitsi_server_url = null;
-    compose.initialize();
-    assert.equal($("#below-compose-content .video_link").visible(), false);
-
-    reset_jquery();
-    stub_out_video_calls();
-    page_params.realm_video_chat_provider =
-        page_params.realm_available_video_chat_providers.jitsi_meet.id;
-    page_params.jitsi_server_url = "https://meet.jit.si";
-    compose.initialize();
-    assert.equal($("#below-compose-content .video_link").visible(), true);
 });
 
 MockDate.reset();
